+++ features/logistic/localizationchooseaddress/src/main/java/com/tokopedia/localizationchooseaddress/domain/model/ChosenAddressModel.kt
@@ -5,7 +5,7 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class ChosenAddressModel (
-        var addressId: Int = 0,
+        var addressId: Long = 0,
         var receiverName: String = "",
         var addressName: String = "",
         var districtId: Int = 0,
