+++ features/logistic/localizationchooseaddress/src/main/java/com/tokopedia/localizationchooseaddress/domain/response/GetChosenAddressListResponse.kt
@@ -20,7 +20,7 @@ data class GetChosenAddressListResponse(
 
 data class ChosenAddressListResponse(
         @SerializedName("addr_id")
-        var addressId: Int = 0,
+        var addressId: Long = 0,
         @SerializedName("receiver_name")
         var receiverName: String = "",
         @SerializedName("addr_name")
