+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/data/model/CustomProductLogisticModel.kt
@@ -18,7 +18,7 @@ data class ShipperListCPLModel(
 )
 
 data class ShipperCPLModel(
-    var shipperId: Int = 0,
+    var shipperId: Long = 0,
     var shipperName: String = "",
     var logo: String = "",
     var shipperProduct: List<ShipperProductCPLModel> = listOf(),
@@ -26,7 +26,7 @@ data class ShipperCPLModel(
 )
 
 data class ShipperProductCPLModel(
-    var shipperProductId: Int = 0,
+    var shipperProductId: Long = 0,
     var shipperProductName: String = "",
     var uiHidden: Boolean = false,
     var isActive: Boolean = false
