+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/data/response/GetDefaultAddressResponse.kt
@@ -14,7 +14,7 @@ data class GetDefaultAddress(
 
 data class DefaultAddressData(
         @SerializedName("addr_id")
-        var addressId: Int = 0,
+        var addressId: Long = 0,
         @SerializedName("receiver_name")
         var receiverName: String = "",
         @SerializedName("addr_name")
