+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/data/response/GetDistrictDetailsResponse.kt
@@ -25,16 +25,16 @@ data class DistrictItem(
         val cityName: String = "",
 
         @SerializedName("province_id")
-        val provinceId: Int = 0,
+        val provinceId: Long = 0,
 
         @SerializedName("district_id")
-        val districtId: Int = 0,
+        val districtId: Long = 0,
 
         @SerializedName("zip_code")
         val zipCode: List<String> = emptyList(),
 
         @SerializedName("city_id")
-        val cityId: Int = 0,
+        val cityId: Long = 0,
 
         @SerializedName("province_name")
         val provinceName: String = ""
