+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/data/response/shippingeditor/GetShipperTickerResponse.kt
@@ -35,16 +35,16 @@ data class HeaderTicker(
         @SerializedName("url_link")
         var urlLink: String = "",
         @SerializedName("warehouse_ids")
-        var warehouseIds: List<Int> = emptyList(),
+        var warehouseIds: List<Long> = emptyList(),
         @SerializedName("is_active")
         var isActive: Boolean = false
 )
 
 data class CourierTicker(
         @SerializedName("shipper_id")
-        var shipperId: Int = -1,
+        var shipperId: Long = -1,
         @SerializedName("warehouse_ids")
-        var warehouseIds: List<Int>? = emptyList(),
+        var warehouseIds: List<Long>? = emptyList(),
         @SerializedName("ticker_state")
         var tickerState: Int = -1,
         @SerializedName("is_available")
@@ -55,26 +55,26 @@ data class CourierTicker(
 
 data class ShipperProductTicker(
         @SerializedName("shipper_product_id")
-        var shipperProductId: Int = -1,
+        var shipperProductId: Long = -1,
         @SerializedName("is_available")
         var isAvailable: Boolean = false
 )
 
 data class Warehouses(
         @SerializedName("warehouse_id")
-        var warehouseId: Int = 0,
+        var warehouseId: Long = 0,
         @SerializedName("warehouse_name")
         var warehouseName: String = "",
         @SerializedName("district_id")
-        var districtId: Int = 0,
+        var districtId: Long = 0,
         @SerializedName("district_name")
         var districtName: String = "",
         @SerializedName("city_id")
-        var cityId: Int = 0,
+        var cityId: Long = 0,
         @SerializedName("city_name")
         var cityName: String = "",
         @SerializedName("province_id")
-        var provinceId: Int = 0,
+        var provinceId: Long = 0,
         @SerializedName("province_name")
         var provinceName: String = "",
         @SerializedName("status")
